AI/recommend/requirement.txt
requests==2.28.2
pymongo==4.4.0
google-api-python-client==2.83.0
scikit-learn==1.2.2
python-dotenv==1.0.0
